requirements.txt
Django==2.2.4
djangorestframework==3.10.2
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
